% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gene.R
\name{wb_gene}
\alias{wb_gene}
\title{Wormbase gene summary}
\usage{
wb_gene(gene)
}
\arguments{
\item{name}{of C. elegans gene to search for.}
}
\value{
data frame of five columns: Gene name, gene type (protein, rRNA, etc),
genomic location, wormbase gene summary, and references. If the gene is not
found in C. elegans, an empty data frame will be returned.
}
\description{
Using a given gene name, the Wormbase REST API is queried
to create a data frame containing a description of the gene,
and other important information.
}

