% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gene.R
\name{wb_variants}
\alias{wb_variants}
\title{Wormbase gene variants}
\usage{
wb_variants(gene)
}
\arguments{
\item{name}{of C. elegans gene to search for.}
}
\value{
data frame of nine columns: Wormbase variant ID, allele name, isoform ID,
locations (exon, intron, etc), molecular change (deletion, subsitution, etc),
gene effect (missense, nonsense, etc), amino acid change, amino acid position,
and associated strain. All empty fields contain NA as a character. If the gene
is not found in C. elegans, an empty data frame will be returned.
}
\description{
Using a given gene name, the Wormbase REST API is queried to create a data
frame containing information on the different variants of the gene, and the
associated strains.
}

